Bandwidth selection for GTWR

A function for automatic bandwidth selection to calibrate a GTWR model

bw.gtwr(formula, data, obs.tv, approach="CV",kernel="bisquare",adaptive=FALSE, 
        p=2, theta=0, longlat=F,lamda=0.05,t.units = "auto",ksi=0, st.dMat,
        verbose=T)

Arguments

formula

Regression model formula of a formula object

data

a Spatial*DataFrame, i.e. SpatialPointsDataFrame or SpatialPolygonsDataFrame as defined in package sp, or a sf object defined in package sf

obs.tv

a vector of time tags for each observation, which could be numeric or of POSIXlt class

approach

specified by CV for cross-validation approach or by AIC corrected (AICc) approach

kernel

function chosen as follows:

gaussian: wgt = exp(-.5*(vdist/bw)^2);

exponential: wgt = exp(-vdist/bw);

bisquare: wgt = (1-(vdist/bw)^2)^2 if vdist < bw, wgt=0 otherwise;

tricube: wgt = (1-(vdist/bw)^3)^3 if vdist < bw, wgt=0 otherwise;

boxcar: wgt=1 if dist < bw, wgt=0 otherwise

adaptive

if TRUE calculate an adaptive kernel where the bandwidth (bw) corresponds to the number of nearest neighbours (i.e. adaptive distance); default is FALSE, where a fixed kernel is found (bandwidth is a fixed distance)

p

the power of the Minkowski distance, default is 2, i.e. the Euclidean distance

theta

an angle in radians to rotate the coordinate system, default is 0

longlat

if TRUE, great circle distances will be calculated

lamda

an parameter between 0 and 1 for calculating spatio-temporal distance

t.units

character string to define time unit

ksi

an parameter between 0 and PI for calculating spatio-temporal distance, see details in Wu et al. (2014)

st.dMat

a pre-specified spatio-temporal distance matrix

verbose

logical variable to define whether show the selection procedure

Note

The function is developed according to the articles by Huang et al. (2010) and Wu et al. (2014).

Value

Returns the adaptive or fixed distance bandwidth

References

Huang, B., Wu, B., & Barry, M. (2010). Geographically and temporally weighted regression for modeling spatio-temporal variation in house prices. International Journal of Geographical Information Science, 24, 383-401.

Wu, B., Li, R., & Huang, B. (2014). A geographically and temporally weighted autoregressive model with application to housing prices. International Journal of Geographical Information Science, 28, 1186-1204.

Fotheringham, A. S., Crespo, R., & Yao, J. (2015). Geographical and Temporal Weighted Regression (GTWR). Geographical Analysis, 47, 431-452.

Author

Binbin Lu binbinlu@whu.edu.cn